requirements.txt
fastapi
uvicorn
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
psycopg2-binary
requests
sqlalchemy
email-validator